chore: further refactorings and improvements
