+++ Velaptor/Telemetry/TelemetryService.cs
@@ -2,6 +2,7 @@
 // Copyright (c) KinsonDigital. All rights reserved.
 // </copyright>
 
+// ReSharper disable once GrammarMistakeInComment
 namespace Velaptor.Telemetry;
 
 #if ENABLE_TELEMETRY
@@ -9,7 +10,6 @@ using System.Globalization;
 using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 using System.Text.Json;
-using System.IO;
 using System;
 using Services;
 #endif
@@ -20,10 +20,8 @@ using System.Diagnostics.CodeAnalysis;
 [ExcludeFromCodeCoverage(Justification = "Telemetry code is challenging to test and provides minimal value to cover with unit tests.")]
 internal class TelemetryService : ITelemetryService
 {
-    // ReSharper disable once NotAccessedField.Local
-    private readonly ITelemetryClient telemetryClient;
-
 #if ENABLE_TELEMETRY
+    private readonly ITelemetryClient telemetryClient;
     private readonly IAppService appService;
     private readonly JsonSerializerOptions serializeOptions = new ()
     {
@@ -46,8 +44,9 @@ internal class TelemetryService : ITelemetryService
     /// <summary>
     /// Initializes a new instance of the <see cref="TelemetryService"/> class.
     /// </summary>
-    /// <param name="telemetryClient">Communicates with the telemetry server.</param>
-    public TelemetryService(ITelemetryClient telemetryClient) => this.telemetryClient = telemetryClient;
+    public TelemetryService()
+    {
+    }
 #endif
 
     /// <inheritdoc/>
@@ -57,6 +56,10 @@ internal class TelemetryService : ITelemetryService
         // Set by anyone (maintainer or game developer) to suppress telemetry.
         // Game developers who do not want telemetry tracked should set this env var to 1 or true.
         var optIn = (Environment.GetEnvironmentVariable("VELAPTOR_OPT_IN_TELEMETRY") ?? string.Empty).ToLower();
+
+        Console.WriteLine($"Opt In State: {optIn}");
+        Console.WriteLine($"Consumer Is Debug: {this.appService.ConsumerIsDebug}");
+        Console.WriteLine($"In Dev Environment: {this.appService.InDevelopmentEnvironment}");
 
         // Has the user opted into telemetry?
         if (string.IsNullOrEmpty(optIn) || optIn is "0" or "false")
@@ -74,7 +77,7 @@ internal class TelemetryService : ITelemetryService
         // Only send telemetry when running in a developer environment.
         // A published game distributed to players will not be running in
         // a game development environment.
-        if (IsInDeveloperEnvironment())
+        if (this.appService.InDevelopmentEnvironment)
         {
             ShowEnabledTelemetryMsg();
 
@@ -105,69 +108,6 @@ internal class TelemetryService : ITelemetryService
 
 #if ENABLE_TELEMETRY
     /// <summary>
-    /// Returns <c>true</c> if the process appears to be running inside a developer's project.
-    /// </summary>
-    /// <remarks>
-    /// <para>
-    /// Primary signal: walk ancestor directories of the running executable looking for
-    /// <c>*.csproj</c> or <c>*.sln</c> files.  These only exist in a development
-    /// environment — a game that has been published and distributed to players will never
-    /// have project files anywhere in its directory tree.
-    /// </para>
-    /// <para>
-    /// Fallback signal: check whether the .NET SDK is installed on this machine.
-    /// Developers must have the SDK installed to build games; players typically only have
-    /// the .NET Runtime or run a self-contained game bundle.
-    /// </para>
-    /// </remarks>
-    private static bool IsInDeveloperEnvironment()
-    {
-        try
-        {
-            var processPath = Environment.ProcessPath ?? string.Empty;
-            var sep = Path.DirectorySeparatorChar;
-
-            // Fast exit path: .NET always builds to bin/Debug/ or bin/Release/.
-            // A distributed game will never have these in its path.
-            if (processPath.Contains($"{sep}bin{sep}Debug{sep}", StringComparison.OrdinalIgnoreCase) ||
-                processPath.Contains($"{sep}bin{sep}Release{sep}", StringComparison.OrdinalIgnoreCase))
-            {
-                return true;
-            }
-
-            var executableDir = Path.GetDirectoryName(processPath) ?? string.Empty;
-            var root = Path.GetPathRoot(executableDir) ?? string.Empty;
-            var dir = executableDir;
-
-            while (!string.IsNullOrEmpty(dir) && !string.Equals(dir, root, StringComparison.OrdinalIgnoreCase))
-            {
-                if (Directory.GetFiles(dir, "*.csproj", SearchOption.TopDirectoryOnly).Length > 0 ||
-                    Directory.GetFiles(dir, "*.sln", SearchOption.TopDirectoryOnly).Length > 0)
-                {
-                    return true;
-                }
-
-                var parent = Path.GetDirectoryName(dir);
-                if (parent is null || string.Equals(parent, dir, StringComparison.OrdinalIgnoreCase))
-                {
-                    break;
-                }
-
-                dir = parent;
-            }
-
-            // No project files found anywhere in the directory tree — this is a distributed game, not a development run.
-            return false;
-        }
-        catch
-        {
-            return true;
-        }
-    }
-
-    // TODO: update url to the learn more about telemetry docs
-
-    /// <summary>
     /// Shows the message that explains the telemetry feature.
     /// </summary>
     private void ShowEnabledTelemetryMsg()
@@ -182,7 +122,7 @@ internal class TelemetryService : ITelemetryService
 
                           Velaptor Telemetry: Enabled
                           ───────────────────────────────────────────────────────────────────────────────────
-                            You have opted into to telemetry, and Velaptor is collecting
+                            You have opted in to telemetry, and Velaptor is collecting
                             anonymous usage data to help us understand how the framework
                             is being used to improve future versions.
 
@@ -200,7 +140,7 @@ internal class TelemetryService : ITelemetryService
                             Data is only collected from the developers that use Velaptor.
                             Data is NOT collected in built/compiled games!
 
-                            Learn more: https://github.com/KinsonDigital/Velaptor/blob/main/docs/telemetry.md
+                            Learn more: https://docs.velaptor.io/telemetry.md
 
                             To opt out, set this environment variable:
                                  VELAPTOR_OPT_IN_TELEMETRY=0 or VELAPTOR_OPT_IN_TELEMETRY=false 
@@ -248,7 +188,7 @@ internal class TelemetryService : ITelemetryService
 
                             To opt in, add the environment variable VELAPTOR_OPT_IN_TELEMETRY=1 or VELAPTOR_OPT_IN_TELEMETRY=true.
 
-                            Learn more: https://github.com/KinsonDigital/Velaptor/blob/main/docs/telemetry.md
+                            Learn more: https://docs.velaptor.io/telemetry.md
                           ────────────────────────────────────────────────────────────────────────────────────────────────────────
 
                           """);
